refactor(inference): extract bench metrics tests to sibling file
Move the single inline `#[cfg(test)] mod tests` block out of
`primer-inference/src/bench/metrics.rs` (636 -> 327 lines) into a sibling
`bench/metrics/tests.rs` (308 lines). The test module is a descendant, so
`use super::*` resolves to the parent module verbatim — no visibility edits,
zero churn. Pure test-only relocation, no behaviour change.

Guard: `cargo test -p primer-inference --lib` -> 199 passed (identical to
baseline). Workspace clippy/fmt/test all green.
